% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gacos_import.R
\name{GACOS.Import}
\alias{GACOS.Import}
\title{Import GACOS product in R}
\usage{
GACOS.Import(rscFile.path, noDataAsNA = FALSE)
}
\arguments{
\item{rscFile.path}{Path of the GACOS .ztd.rsc file}

\item{noDataAsNA}{If true it convert 0 values to NA}
}
\description{
Import GACOS product in R
}
\examples{
library(raster)
library(GInSARCorW)
library(circular)
rscFile.path<-system.file("td","20170317.ztd.rsc",package = "GInSARCorW")
noDataAsNA<-FALSE
GACOS.Import(rscFile.path,noDataAsNA)
}
\author{
Subhadip Datta
}
